+++ src/app/mypage/myProfile/type.ts
